dom/media/mediasource/test/test_FrameSelection_mp4.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: Don't get stuck buffering for too long when we have frames to show</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// This test loads partial video, plays and waits until playback stalls.
// It then loads only 3 frames of a video at higher resolution.

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  ok(true, "Receive a sourceopen event");
  ms.addEventListener("sourceopen", () => ok(false, "No more sourceopen"));
  const sb = ms.addSourceBuffer("video/mp4");
  ok(sb, "Create a SourceBuffer");
  logEvents(v);
  sb.addEventListener("error", e => {
    ok(false, `should not fire ${e.type} event`);
    SimpleTest.finish();
  });
  await fetchAndLoad(sb, "bipbop/bipbop", ["init"], ".mp4");
  const p = once(v, "loadeddata");
  await fetchAndLoad(sb, "bipbop/bipbop", range(1, 3), ".m4s");
  await p;
  is(sb.buffered.length, 1, "continuous range");
  v.play();
  // We have nothing to play, waiting will be fired.
  await waitUntilTime(v, 1.5);
  await fetchAndLoad(sb, "bipbop/bipbop_480_624kbps-video", ["init"], ".mp4");
  sb.timestampOffset = 1.601666; // End of the video track buffered - time of first video sample (0.095).
  sb.appendWindowEnd = 1.796677; // Only allow room for three extra video frames (we need 3 as this video has b-frames).
  await fetchAndLoad(sb, "bipbop/bipbop_480_624kbps-video", ["1"], ".m4s");
  ms.endOfStream();
  await Promise.all([once(ms, "sourceended"), once(v, "playing"), once(v, "ended")]);
  if (v.width, 640, "has proper width");
  if (v.height, 480, "has proper height");
  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
